% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_languages.R
\name{get_languages}
\alias{get_languages}
\title{Get all languages available on the ODS platform}
\usage{
get_languages(page_elements, verbose = TRUE)
}
\arguments{
\item{page_elements}{List object from get_page(). Contains the title, description, template, language, content, tags and level of restriction of the page.}

\item{verbose}{A logical TRUE/FALSE indicating if the function should output a message listing the languages available on the platform.}
}
\value{
A vector of characters.
}
\description{
Retrieve all languages available on the ODS platform from the information contained in page_elements.
}
\examples{
# Temporary directory for reproducible example
dir_tmp <- tempfile(pattern = "proj-")
dir.create(dir_tmp)

file.copy(from = system.file("examples/page_elements_example", package = "odsrmd"), to = dir_tmp)
page_elements <- readRDS(paste0(dir_tmp, "/page_elements_example"))

all_languages <- get_languages(page_elements)
}
