% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ods_page.R
\name{get_ods_page}
\alias{get_ods_page}
\title{Get ODS page}
\usage{
get_ods_page(page_slug)
}
\arguments{
\item{page_slug}{Slug of the page.}
}
\value{
A list object containing the information from the page.
}
\description{
Get the content of an ODS page.
}
\examples{
get_ods_page(page_slug = "test")

}
