% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_ods_page.R
\name{put_ods_page}
\alias{put_ods_page}
\title{Put ODS page}
\usage{
put_ods_page(page_slug, json_to_send)
}
\arguments{
\item{page_slug}{Slug of the page.}

\item{json_to_send}{JSON object that will be send to the Opendatasoft portal and which information will be displayed on the page corresponding to the page slug.}
}
\value{
If request is successful (i.e. the request was successfully performed and a response with HTTP status code <400 was received), an HTTP response; otherwise throws an error.
}
\description{
Send JSON element to an ODS page.
}
\examples{

# Temporary directory for reproducible example
dir_tmp <- tempfile(pattern = "proj-")
dir.create(dir_tmp)

file.copy(from = system.file("data/example_rmd.Rmd", package = "odsrmd"), to=dir_tmp)
# browseURL(dir_tmp)
path <- paste0(dir_tmp, "/example_rmd.Rmd")

body_and_style <- get_body_and_style(path)


page_elements <- get_ods_page(page_slug = "test")

json_to_send <- create_json(page_elements, body_and_style, language = NULL, 
                        title = NULL, description = NULL, template = NULL, 
                        tags = NULL, restricted = NULL)
page_slug <- "test"

put_ods_page(page_slug, json_to_send)


}
