% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_json.R
\name{create_json}
\alias{create_json}
\title{Create JSON element to send to ODS platform.}
\usage{
create_json(
  page_elements,
  body_and_style,
  title = NULL,
  description = NULL,
  template = NULL,
  chosen_languages = "all",
  tags = NULL,
  restricted = NULL
)
}
\arguments{
\item{page_elements}{List object from get_page(). Contains the title, description, template, language, content, tags and level of restriction of the page. By default, if no value is provided to title, description, template, language, tags and restricted, create_json() takes the values available in page_elements.}

\item{body_and_style}{List object containing body and style of knitted Rmd document coming from get_body_and_style().}

\item{title}{A character string specifying the title of the page. If not null, overwrites title from page_elements.}

\item{description}{A character string containing the description of the page. If not null, overwrites description from page_elements.}

\item{template}{The template of the page. Default to "custom.html". If not null, overwrites template from page_elements.}

\item{chosen_languages}{A character vector of languages in which the page should be updated. Default to "all", meaning all languages available on the platform.}

\item{tags}{A list object of character strings containing the tags describing the page. If not null, overwrites tags from page_elements.}

\item{restricted}{A boolean TRUE/FALSE indicating the reading status of the page. TRUE will make the page "public" whereas FALSE will keep its access restricted to users who where granted the permission. If not null, overwrites reading status from page_elements.}
}
\value{
a JSON element
}
\description{
Combine information coming from the Opendatasoft page (with get_ods_page()) and the Rmarkdown document (with get_body_and_style()) into a JSON object ready to be sent back to the Opendatasoft platform.
}
\examples{
# Temporary directory for reproducible example
dir_tmp <- tempfile(pattern = "proj-")
dir.create(dir_tmp)

file.copy(from = system.file("examples/example_rmd.Rmd", package = "odsrmd"), to = dir_tmp)
# browseURL(dir_tmp)
path <- paste0(dir_tmp, "/example_rmd.Rmd")
page_slug <- "testthat-odsrmd"


body_and_style <- get_body_and_style(path)

page_elements <- get_ods_page(page_slug)

json_to_send <- create_json(page_elements, body_and_style,
  chosen_languages = "all",
  title = NULL, description = NULL, template = NULL,
  tags = NULL, restricted = NULL
)
}
