% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_json.R
\name{create_json}
\alias{create_json}
\title{Create JSON element to send to ODS portal}
\usage{
create_json(
  page_elements,
  body_and_style,
  title = NULL,
  description = NULL,
  template = NULL,
  language = NULL,
  tags = NULL,
  restricted = NULL
)
}
\arguments{
\item{page_elements}{List object coming from get_page(). Contains the title, description, template, language, content, tags and level of restriction of the page. By default, if no value is provided to title, description, template, language, tags and restricted, create_json() takes the values available in page_elements.}

\item{body_and_style}{List object containing body and style of knitted Rmd document coming from get_body_and_css().}

\item{title}{A character string specifying the title of the page.}

\item{description}{A character string containing the description of the page.}

\item{template}{The template of the page. Default to "custom.html".}

\item{language}{The language in which you would like to post the page. It should be one of the languages available on the Opendatasoft portal hosting the page.}

\item{tags}{A list object of character strings containing the tags describing the page.}

\item{restricted}{A boolean TRUE/FALSE indicating the reading status of the page. TRUE will make the page "public" whereas FALSE will keep its access restricted to users who where granted the permission.}
}
\value{
a JSON element
}
\description{
Modify list object coming from get_page and returns a JSON object.
}
\examples{
# Temporary directory for reproducible example
dir_tmp <- tempfile(pattern = "proj-")
dir.create(dir_tmp)

file.copy(from = system.file("data/example_rmd.Rmd", package = "odsrmd"), to=dir_tmp)
# browseURL(dir_tmp)
path <- paste0(dir_tmp, "/example_rmd.Rmd")

body_and_style <- get_body_and_style(path)


page_elements <- get_ods_page(page_slug = "test")

create_json(page_elements, body_and_style, language = NULL, 
                        title = NULL, description = NULL, template = NULL, 
                        tags = NULL, restricted = NULL)
}
