% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_body_and_style.R
\name{get_body_and_style}
\alias{get_body_and_style}
\title{Get body and style}
\usage{
get_body_and_style(path)
}
\arguments{
\item{path}{Path to the Rmd file.}
}
\value{
A named list with two elements : body and style.
}
\description{
Render Rmarkdown file as html file and extract the html nodes corresponding to body and style without the script nodes.
}
\examples{
# Temporary directory for reproducible example
dir_tmp <- tempfile(pattern = "proj-")
dir.create(dir_tmp)

file.copy(from = system.file("examples/example_rmd.Rmd", package = "odsrmd"), to=dir_tmp)
# browseURL(dir_tmp)
path <- paste0(dir_tmp, "/example_rmd.Rmd")

body_and_style <- get_body_and_style(path)
}
