% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_languages.R
\name{get_languages}
\alias{get_languages}
\title{Get all languages available on the ODS platform}
\usage{
get_languages(page_elements, verbose = TRUE)
}
\arguments{
\item{page_elements}{List object from get_page(). Contains the title, description, template, language, content, tags and level of restriction of the page.}

\item{verbose}{A boolean TRUE/FALSE indicating if the function should output a message listing the languages available on the platform.}
}
\value{
A vector of characters.
}
\description{
Retrieve all languages available on the ODS platform from the information contained in page_elements.
}
\examples{
page_slug <- "testthat-odsrmd"
page_elements <- get_ods_page(page_slug) 

all_languages <- get_languages(page_elements)
}
