% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ods_page.R
\name{get_ods_page}
\alias{get_ods_page}
\title{Get ODS page}
\usage{
get_ods_page(page_slug)
}
\arguments{
\item{page_slug}{Slug of the page (in the backoffice of the platform -> Pages -> New or existing page -> Properties -> Page URL).}
}
\value{
A list object containing the information from the page (title, description, tags...).
}
\description{
Get the content of a classic page from an Opendatasoft platform.
}
\examples{
page_slug <- "testthat-odsrmd" 
page_elements <- get_ods_page(page_slug) 

}
