% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_available.R
\name{is_available}
\alias{is_available}
\title{Check language availability}
\usage{
is_available(language, page_elements)
}
\arguments{
\item{language}{Character string.}

\item{page_elements}{List object from get_page(). Contains the title, description, template, language, content, tags and level of restriction of the page.}
}
\description{
Check if a given language is available on the ODS platform from the list of languages contained in page_elements.
}
\examples{
page_slug <- "testthat-odsrmd" 
page_elements <- get_ods_page(page_slug) 

is_available("EN", page_elements)
}
